pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-control-plane"
version = "0.1.0"
description = "Operator-grade control plane for AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "httpx>=0.27,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "alembic>=1.14,<2.0",
  "PyJWT>=2.9,<3.0"
]

[project.scripts]
agent-control-plane-worker = "agent_control_plane.worker:run_forever"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
